ansible: move infrastructure management to ansible
This gives us repeatable and open-source infrastructure.

This commit is a megacommit and this error of mine is a grave one.
Progress has regressed multiple times due to accidental deletion and a
lack of version control. I deeply regret not committing throughout this
process.

This is not the final state of this tree, roles are not yet as reusable
as I'd like them to be, and docs are not up to the standard I'd like
them to be. These issues will be addressed in the future, as our growing
infrastructure demands more convenience for me.